MVVM + Clean Architecture (исправленное)
