--- app/src/main/java/com/example/architecturebase/PostContract.kt
@@ -1,9 +1,0 @@
-package com.example.architecturebase
-
-import com.example.architecturebase.network.IPostApi
-
-interface PostContract {
-    interface IPostPresenter {
-        fun getPostsApi(): IPostApi
-    }
-}
