--- app/src/main/java/com/example/architecturebase/PostPresenter.kt
@@ -1,9 +1,0 @@
-package com.example.architecturebase
-
-import com.example.architecturebase.network.IPostApi
-
-class PostPresenter : PostContract.IPostPresenter {
-
-    override fun getPostsApi(): IPostApi = RetrofitInstance.getRetrofit()
-
-}
