+++ app/src/main/java/com/example/architecturebase/domain/repos/NewsRepo.kt
@@ -1,0 +1,5 @@
+package com.example.architecturebase.domain.repos
+
+interface NewsRepo {
+    fun getNews()
+}
