+++ app/src/main/java/com/example/architecturebase/presentation/MainActivity.kt
@@ -1,12 +1,13 @@
-package com.example.architecturebase
+package com.example.architecturebase.presentation
 
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.architecturebase.adapter.MainAdapter
+import com.example.architecturebase.presentation.adapter.MainAdapter
 import com.example.architecturebase.databinding.ActivityMainBinding
+import com.example.architecturebase.presentation.viewmodel.PostsViewModel
 
 class MainActivity : AppCompatActivity() {
