+++ app/src/main/java/com/example/architecturebase/presentation/adapter/MainAdapter.kt
@@ -1,10 +1,10 @@
-package com.example.architecturebase.adapter
+package com.example.architecturebase.presentation.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.architecturebase.databinding.PostViewBinding
-import com.example.architecturebase.network.model.Post
+import com.example.architecturebase.domain.entities.Post
 
 class MainAdapter : RecyclerView.Adapter<MainAdapter.Holder>() {
