+++ app/src/main/java/com/example/architecturebase/presentation/contract/PostContract.kt
@@ -1,0 +1,9 @@
+package com.example.architecturebase.presentation.contract
+
+import com.example.architecturebase.data.api.IPostApi
+
+interface PostContract {
+    interface IPostPresenter {
+        fun getNews()
+    }
+}
